introduce SymbolsDatabase
